Check file auto.cnf file existence (#50)

+++ init-script/run.sh
@@ -5,6 +5,8 @@ rm -rf /run-scripts/*
 cp /tmp/scripts/* /scripts
 if [[ "$PITR_RESTORE" == "true" ]]; then
   if [[ "$HOSTNAME" != *"-0" ]]; then
-    rm /var/lib/mysql/auto.cnf
+    if [[ -f /var/lib/mysql/auto.cnf ]]; then
+       rm /var/lib/mysql/auto.cnf
+    fi
   fi
 fi
